+++ Sadie.Shared/HPoint.cs
@@ -4,9 +4,9 @@ public class HPoint
 {
     public int X { get; }
     public int Y { get; }
-    public int Z { get; }
+    public double Z { get; }
     
-    public HPoint(int x, int y, int z)
+    public HPoint(int x, int y, double z)
     {
         X = x;
         Y = y;
